+++ hooks/rancher-clusters-to-argocd-clusters.sh
@@ -3,6 +3,8 @@
 set -e
 #set -x
 
+: ${ENABLE_HOOK_RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS:="false"}
+
 # meant to be configured externally at launch time
 : ${ENVIRONMENT_ID:=""}
 : ${REGION_ID:=""}
@@ -14,6 +16,12 @@ set -e
 : ${ARGOCD_NAMESPACE:=argocd}
 
 : ${RANCHER_URI:="https://rancher.cattle-system"}
+
+: ${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_EXCLUDE_REGEX:=""}
+: ${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_INCLUDE_REGEX:=""}
+
+: ${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_REMOVE_TOKEN_TTL:="false"}
+
 # K8S_CA_DATA takes precedence over these
 # dynamically pull the CA data from cluster secret
 : ${RANCHER_CA_SECRET_NAME:=""}
@@ -30,8 +38,10 @@ set -e
 : ${K8S_CA_DATA:=""}
 : ${K8S_INSECURE:="false"}
 
+# https://github.com/flant/shell-operator/issues/726
 if [[ $1 == "--config" ]]; then
-  cat <<EOF
+  if [[ "${ENABLE_HOOK_RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS}" == "true" ]]; then
+    cat <<EOF
 configVersion: v1
 kubernetes:
 - apiVersion: management.cattle.io/v3
@@ -42,6 +52,15 @@ schedule:
   crontab: "*/15 * * * *"
   allowFailure: true
 EOF
+  else
+    cat <<EOF
+configVersion: v1
+settings:
+  executionMinInterval: 1s
+  executionBurst: 1
+EOF
+  fi
+
   exit 0
 fi
 
@@ -89,6 +108,11 @@ if [[ -z "${K8S_TOKEN}" ]]; then
     exit 1
   fi
 
+  if [[ "${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_REMOVE_TOKEN_TTL}" == "true" ]]; then
+    echo "removing token ttl"
+    kubectl patch tokens.management.cattle.io "${tokenResourceName}" --type='merge' -p '{ "expiresAt": "", "ttl": 0 }'
+  fi
+
   K8S_TOKEN="${tokenResourceName}:${userToken}"
   echo "properly fetched bearerToken from rancher crds"
 fi
@@ -98,20 +122,37 @@ if [[ -z "${K8S_TOKEN}" ]]; then
   exit 1
 fi
 
+RANCHER_CLUSTERS=$(kubectl get clusters.management.cattle.io -o json)
+
 # iterate rancher clusters and create corresponding argocd clusters
-kubectl get clusters.management.cattle.io -o json | jq -crM '.items[]' | while read -r cluster; do
+echo $RANCHER_CLUSTERS | jq -crM '.items[]' | while read -r cluster; do
   # removing this label so rancher does not remove the secret immediately thinking it owns the secret
   cluster=$(echo "${cluster}" | jq -crM 'del(.metadata.labels."objectset.rio.cattle.io/hash")')
   clusterResourceName=$(echo "${cluster}" | jq -crM '.metadata.name')
-  echo "handling cluster: ${clusterResourceName}"
+  clusterDisplayName=$(echo "${cluster}" | jq -crM '.spec.displayName')
+  echo "cluster ${clusterResourceName} display name: ${clusterDisplayName}"
 
   if [[ -z "${clusterResourceName}" ]]; then
     echo "empty cluster, moving on"
     continue
   fi
 
-  displayName=$(echo "${cluster}" | jq -crM '.spec.displayName')
-  echo "cluster ${clusterResourceName} display name: ${displayName}"
+  # cluster exclude filtering
+  if [[ -n "${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_EXCLUDE_REGEX}" ]]; then
+    if [[ "${clusterDisplayName}" =~ ${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_EXCLUDE_REGEX} ]]; then
+      echo "ignoring cluster ${clusterDisplayName} due to exclude filtering"
+      continue
+    fi
+  fi
+
+  # cluster include filtering
+  if [[ -n "${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_INCLUDE_REGEX}" ]]; then
+    if ! [[ "${clusterDisplayName}" =~ ${RANCHER_CLUSTERS_TO_ARGOCD_CLUSTERS_CLUSTER_NAME_INCLUDE_REGEX} ]]; then
+      echo "ignoring cluster ${clusterDisplayName} due to include filtering"
+      continue
+    fi
+  fi
+
   # this is kube-ca
   caCert=$(echo "${cluster}" | jq -crM '.status.caCert')
   clusterLabels=$(echo "${cluster}" | yq eval '.metadata.labels' - -P | sed "s/^/    /g")
@@ -125,13 +166,13 @@ metadata:
   labels:
 ${clusterLabels}
     argocd.argoproj.io/secret-type: cluster
-    clusterId: "${displayName}"
+    clusterId: "${clusterDisplayName}"
     environmentId: "${ENVIRONMENT_ID}"
     regionId: "${REGION_ID}"
     rancherImported: "true"
 type: Opaque
 stringData:
-  name: "${SERVER_NAME_PREFIX}${displayName}${SERVER_NAME_SUFFIX}"
+  name: "${SERVER_NAME_PREFIX}${clusterDisplayName}${SERVER_NAME_SUFFIX}"
   server: "${RANCHER_URI}/k8s/clusters/${clusterResourceName}"
   config: |
     {